Change framing byte properties from int to byte, remove Range attributes and casts
Agent-Logs-Url: https://github.com/FrendsPlatform/Frends.Mllp/sessions/7787dbb4-1c70-48db-8d67-612f2be532da

+++ Frends.Mllp.Receive/Frends.Mllp.Receive/Frends.Mllp.Receive.cs
@@ -156,9 +156,9 @@ public static class Mllp
                 services.AddSingleton(encoding);
                 services.AddSingleton(new MllpFramingBytes
                 {
-                    StartBlock = (byte)options.StartBlockByte,
-                    EndBlock = (byte)options.EndBlockByte,
-                    CarriageReturn = (byte)options.CarriageReturnByte,
+                    StartBlock = options.StartBlockByte,
+                    EndBlock = options.EndBlockByte,
+                    CarriageReturn = options.CarriageReturnByte,
                 });
             })
             .UsePackageHandler(async (session, package) =>
